finished generalization and connected to button
